pkg/endpoint: move runLabelsResolver to identity.go
Signed-off by: Ian Vernon <ian@cilium.io>
