+++ pkg/endpoint/endpoint.go
@@ -1603,18 +1603,6 @@ func (e *Endpoint) UpdateLabels(owner Owner, identityLabels, infoLabels labels.L
 	}
 }
 
-func (e *Endpoint) runLabelsResolver(owner Owner, myChangeRev int) {
-	ctrlName := fmt.Sprintf("resolve-identity-%d", e.ID)
-	e.controllers.UpdateController(ctrlName,
-		controller.ControllerParams{
-			DoFunc: func() error {
-				return e.identityLabelsChanged(owner, myChangeRev)
-			},
-			RunInterval: time.Duration(5) * time.Minute,
-		},
-	)
-}
-
 func (e *Endpoint) identityLabelsChanged(owner Owner, myChangeRev int) error {
 	e.Mutex.RLock()
 	newLabels := e.OpLabels.IdentityLabels()
